refactor: 코드 리뷰 반영 리팩토링
