--- src/main/kotlin/com/stepbookstep/server/auth/dto/ReissueRequest.kt
@@ -1,6 +1,0 @@
-package com.stepbookstep.server.auth.dto
-
-data class ReissueRequest(
-    @field:jakarta.validation.constraints.NotBlank
-    val refreshToken: String
-)
