+++ src/main/kotlin/com/stepbookstep/server/domain/auth/application/AuthService.kt
@@ -1,8 +1,9 @@
-package com.stepbookstep.server.auth
+package com.stepbookstep.server.domain.auth.application
 
-import com.stepbookstep.server.auth.dto.KakaoLoginRequest
-import com.stepbookstep.server.auth.dto.KakaoLoginResponse
-import com.stepbookstep.server.auth.dto.LogoutRequest
+import com.stepbookstep.server.domain.user.application.UserService
+import com.stepbookstep.server.domain.auth.application.dto.KakaoLoginRequest
+import com.stepbookstep.server.domain.auth.application.dto.KakaoLoginResponse
+import com.stepbookstep.server.domain.auth.application.dto.LogoutRequest
 import com.stepbookstep.server.external.kakao.KakaoApiClient
 import com.stepbookstep.server.external.kakao.KakaoUserMeResponse
 import com.stepbookstep.server.global.response.CustomException
@@ -11,10 +12,10 @@ import com.stepbookstep.server.security.jwt.JwtProvider
 import com.stepbookstep.server.security.jwt.TokenType
 import com.stepbookstep.server.security.token.RefreshTokenService
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
 
 /**
  * 카카오 로그인 비즈니스 로직 (로그인/재발급/로그아웃)을 담당
+ * - 외부 API 호출과 DB 트랜잭션을 분리하여 성능을 최적화
  */
 @Service
 class AuthService(
@@ -24,31 +25,28 @@ class AuthService(
     private val refreshTokenService: RefreshTokenService,
 ) {
 
-    @Transactional
+    /**
+     * 카카오 로그인 전체 흐름 제어 (트랜잭션 없음)
+     */
     fun kakaoLogin(request: KakaoLoginRequest): KakaoLoginResponse {
         val kakaoAccessToken = request.socialToken.trim()
         if (kakaoAccessToken.isBlank()) {
             throw CustomException(ErrorCode.INVALID_INPUT)
         }
 
-        // 1) 카카오 사용자 정보 조회
+        // 1) 카카오 사용자 정보 조회 (외부 API 호출 - Network I/O)
         val kakaoMe: KakaoUserMeResponse = kakaoApiClient.getMe(kakaoAccessToken)
         val providerUserId = kakaoMe.id.toString()
-        val nicknameFromKakao = kakaoMe.nickname
+        val nicknameFromKakao = kakaoMe.nickname ?: "사용자"
 
-        // 2) 유저 조회/생성 + 신규 여부 판단
-        val existingUser = userService.findByKakaoProviderUserId(providerUserId)
-        val user = existingUser ?: userService.createKakaoUser(
-            providerUserId = providerUserId,
-            nickname = nicknameFromKakao ?: "사용자"
-        )
-        val isNewUser = (existingUser == null)
+        // 2) 유저 조회/생성 (DB Transaction - UserService 내부에서 처리)
+        val (user, isNewUser) = userService.getOrCreateKakaoUser(providerUserId, nicknameFromKakao)
 
-        // 3) JWT 발급
+        // 3) JWT 발급 (CPU 연산)
         val accessToken = jwtProvider.createToken(user.id, TokenType.ACCESS)
         val refreshToken = jwtProvider.createToken(user.id, TokenType.REFRESH)
 
-        // 4) refresh token 저장 (SHA-256 해시는 RefreshTokenService 내부에서 처리)
+        // 4) Refresh Token 저장 (DB Transaction - RefreshTokenService 내부에서 처리)
         refreshTokenService.save(userId = user.id, refreshToken = refreshToken)
 
         return KakaoLoginResponse(
