+++ src/main/kotlin/com/stepbookstep/server/domain/auth/application/dto/ReissueRequest.kt
@@ -1,0 +1,8 @@
+package com.stepbookstep.server.domain.auth.application.dto
+
+import jakarta.validation.constraints.NotBlank
+
+data class ReissueRequest(
+    @field:NotBlank
+    val refreshToken: String
+)
