+++ src/main/kotlin/com/stepbookstep/server/domain/user/application/UserDeviceService.kt
@@ -1,4 +1,4 @@
-package com.stepbookstep.server.auth
+package com.stepbookstep.server.domain.user.application
 
 /**
  * 사용자 디바이스 정보 활용시 사용되는 인터페이스 (현재 사용 X)
