+++ src/main/kotlin/com/stepbookstep/server/domain/user/domain/User.kt
@@ -1,6 +1,11 @@
-package com.stepbookstep.server.user
+package com.stepbookstep.server.domain.user.domain
 
-import jakarta.persistence.*
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+import jakarta.persistence.Table
 import java.time.OffsetDateTime
 
 @Entity
