+++ src/main/kotlin/com/stepbookstep/server/domain/user/domain/UserRepository.kt
@@ -1,4 +1,4 @@
-package com.stepbookstep.server.user
+package com.stepbookstep.server.domain.user.domain
 
 import org.springframework.data.jpa.repository.JpaRepository
